third_party/WebKit/LayoutTests/virtual/threaded/transitions/transition-end-event-multiple-03-expected.txt

Initiating transitions on various properties of all boxes, with different start times on the transitions.

FAIL - Missing events
[Missing] Property: background-color Target: box2 Elapsed Time: 0.05
[Missing] Property: background-color Target: box3 Elapsed Time: 0.06
[Missing] Property: height Target: box3 Elapsed Time: 0.06
[Missing] Property: height Target: box4 Elapsed Time: 0.03
[Missing] Property: left Target: box1 Elapsed Time: 0.05
[Missing] Property: left Target: box2 Elapsed Time: 0.05
[Missing] Property: left Target: box3 Elapsed Time: 0.06
[Missing] Property: left Target: box4 Elapsed Time: 0.03
[Missing] Property: width Target: box3 Elapsed Time: 0.06
